+++ src/aston/group17/model/Vehicle.java
@@ -85,7 +85,7 @@ public abstract class Vehicle {
 	 
 	 public abstract int timeSpent();
 
-	 public abstract int moneySpent();
+	 public abstract double moneySpent();
